회원가입 및 회원정보 수정 리팩토링
