+++ src/Modals/UserInfo/PersonalitiesModal.tsx
@@ -1,4 +1,4 @@
-import React, {useEffect, useMemo, useRef, useState} from 'react';
+import React, {useEffect, useMemo} from 'react';
 import {
   Text,
   View,
@@ -7,16 +7,15 @@ import {
   Animated,
   ScrollView,
 } from 'react-native';
-import {ResetButton} from '../Components/ResetButton';
-import useStore from '../Store/store';
-import {ModalHeader} from '../Components/ModalHeader';
-import {SCREEN_HEIGHT} from '../Constants/screen';
-import {UpdateButton} from '../Components/UpdateButton';
-import {Personalities} from '../types/types';
-import {getPersonalities} from '../APIs/personality';
-import {PersonalityList} from '../Components/PersonalityList';
+import {ResetButton} from '../../Components/ResetButton';
+import useStore from '../../Store/store';
+import {ModalHeader} from '../../Components/ModalHeader';
+import {SCREEN_HEIGHT} from '../../Constants/screen';
+import {UpdateButton} from '../../Components/UpdateButton';
+import {PersonalityList} from '../../Components/PersonalityList';
 import {useSafeAreaInsets} from 'react-native-safe-area-context';
-import {patchUserInfo} from '../APIs/member';
+import {patchUserInfo} from '../../APIs/member';
+import {usePersonality} from '../../Hooks/UserInfo/usePersonality';
 
 type Props = {
   isModalVisible: boolean;
@@ -24,54 +23,28 @@ type Props = {
 };
 
 export function PersonalitiesModal({isModalVisible, setModalVisible}: Props) {
-  const [personalities, setPersonalities] = useState<Personalities>([]);
-  const [selectedPersonalityIds, setSelectedPersonalityIds] = useState<
-    number[]
-  >([]);
-  const [activateUpdate, setActivateUpdate] = useState(true);
+  const {
+    personalities,
+    selectedPersonalityIds,
+    setSelectedPersonalityIds,
+    selectPersonality,
+    alertOpacity,
+    counter,
+    reset,
+    resetAlert,
+  } = usePersonality();
 
   const {userInfo} = useStore();
 
   const {bottom: SAFE_AREA_BOTTOM} = useSafeAreaInsets();
 
-  const fadeAnim = useRef(new Animated.Value(0)).current;
-
-  const alert = Animated.sequence([
-    Animated.timing(fadeAnim, {
-      toValue: 1,
-      duration: 0,
-      useNativeDriver: true,
-    }),
-    Animated.timing(fadeAnim, {
-      toValue: 0,
-      delay: 2000,
-      useNativeDriver: true,
-    }),
-  ]);
-
-  const counter = useMemo(
-    () => selectedPersonalityIds.length,
+  const disableUpdate = useMemo(
+    () => selectedPersonalityIds.length === 0,
     [selectedPersonalityIds],
   );
 
-  const selectPersonality = (personalityId: number) => {
-    alert.reset();
-    if (
-      counter < 9 &&
-      selectedPersonalityIds.includes(personalityId) === false
-    ) {
-      setSelectedPersonalityIds([...selectedPersonalityIds, personalityId]);
-    } else if (selectedPersonalityIds.includes(personalityId) === true) {
-      setSelectedPersonalityIds(
-        [...selectedPersonalityIds].filter(e => e !== personalityId),
-      );
-    } else {
-      alert.start();
-    }
-  };
-
   const hideModal = () => {
-    alert.reset();
+    resetAlert();
     setModalVisible(false);
   };
 
@@ -90,35 +63,11 @@ export function PersonalitiesModal({isModalVisible, setModalVisible}: Props) {
     }
   };
 
-  const reset = () => {
-    setSelectedPersonalityIds([]);
-  };
-
   useEffect(() => {
-    try {
-      getPersonalities().then(personalityData => {
-        setPersonalities(personalityData);
-      });
-    } catch (error: any) {
-      console.error(error.message);
+    if (userInfo && userInfo.personalityIds) {
+      setSelectedPersonalityIds(userInfo.personalityIds);
     }
-  }, []);
-
-  useEffect(() => {
-    if (isModalVisible) {
-      if (userInfo && userInfo.personalityIds) {
-        setSelectedPersonalityIds(userInfo.personalityIds);
-      }
-    }
-  }, [isModalVisible, userInfo]);
-
-  useEffect(() => {
-    if (counter > 0) {
-      setActivateUpdate(true);
-    } else {
-      setActivateUpdate(false);
-    }
-  }, [counter]);
+  }, [setSelectedPersonalityIds, userInfo]);
 
   return (
     <Modal
@@ -155,12 +104,12 @@ export function PersonalitiesModal({isModalVisible, setModalVisible}: Props) {
               selectedPersonalityIds={selectedPersonalityIds}
             />
           </ScrollView>
-          <Animated.View style={[styles.alert, {opacity: fadeAnim}]}>
+          <Animated.View style={[styles.alert, {opacity: alertOpacity}]}>
             <Text style={styles.alertText}>최대 9개까지만 선택 가능해요!</Text>
           </Animated.View>
 
           <UpdateButton
-            activateUpdate={activateUpdate}
+            activateUpdate={!disableUpdate}
             onPress={updatePersonalites}
           />
         </View>
