+++ src/Screens/SignUp/PersonalityForm.tsx
@@ -1,4 +1,4 @@
-import React, {useEffect, useMemo, useRef, useState} from 'react';
+import React, {useMemo} from 'react';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
 import {
   View,
@@ -16,82 +16,32 @@ import {SCREEN_HEIGHT} from '../../Constants/screen';
 import {NextButton} from '../../Components/NextButton';
 import useStore from '../../Store/store';
 import {ResetButton} from '../../Components/ResetButton';
-import {getPersonalities} from '../../APIs/personality';
-import {Personalities} from '../../types/types';
 import {PersonalityList} from '../../Components/PersonalityList';
+import {usePersonality} from '../../Hooks/UserInfo/usePersonality';
 
 type Props = NativeStackScreenProps<StackParamsList, 'PersonalityForm'>;
 
 export function PersonalityForm({navigation}: Props) {
-  const store = useStore();
-  const [personalities, setPersonalities] = useState<Personalities>([]);
-  const [selectedPersonalityIds, setSelectedPersonalityIds] = useState<
-    number[]
-  >([]);
-  const [activateNext, setActivateNext] = useState<boolean>(false);
+  const {
+    personalities,
+    selectedPersonalityIds,
+    selectPersonality,
+    alertOpacity,
+    counter,
+    reset,
+  } = usePersonality();
 
-  const fadeAnim = useRef(new Animated.Value(0)).current;
-
-  const alert = Animated.sequence([
-    Animated.timing(fadeAnim, {
-      toValue: 1,
-      duration: 0,
-      useNativeDriver: true,
-    }),
-    Animated.timing(fadeAnim, {
-      toValue: 0,
-      delay: 2000,
-      useNativeDriver: true,
-    }),
-  ]);
-
-  const counter = useMemo(
-    () => selectedPersonalityIds.length,
+  const disableNext = useMemo(
+    () => selectedPersonalityIds.length === 0,
     [selectedPersonalityIds],
   );
 
-  const selectPersonality = (personalityId: number) => {
-    alert.reset();
-    if (
-      counter < 9 &&
-      selectedPersonalityIds.includes(personalityId) === false
-    ) {
-      setSelectedPersonalityIds([...selectedPersonalityIds, personalityId]);
-    } else if (selectedPersonalityIds.includes(personalityId) === true) {
-      setSelectedPersonalityIds(
-        [...selectedPersonalityIds].filter(e => e !== personalityId),
-      );
-    } else {
-      alert.start();
-    }
-  };
+  const store = useStore();
 
   const goToLocationForm = () => {
     store.setPersonalityIds(selectedPersonalityIds);
     navigation.navigate('LocationForm');
   };
-
-  const reset = () => {
-    setSelectedPersonalityIds([]);
-  };
-
-  useEffect(() => {
-    try {
-      getPersonalities().then(personalityData => {
-        setPersonalities(personalityData);
-      });
-    } catch (error: any) {
-      console.error(error.message);
-    }
-  }, []);
-
-  useEffect(() => {
-    if (counter > 0) {
-      setActivateNext(true);
-    } else {
-      setActivateNext(false);
-    }
-  }, [counter]);
 
   return (
     <LinearGradient colors={['#ffccee', 'white', 'white', 'white', '#ffffcc']}>
@@ -120,11 +70,11 @@ export function PersonalityForm({navigation}: Props) {
           />
         </ScrollView>
         <View style={styles.alertBox}>
-          <Animated.View style={{opacity: fadeAnim}}>
+          <Animated.View style={{opacity: alertOpacity}}>
             <Text style={styles.alertText}>최대 9개까지만 선택 가능해요!</Text>
           </Animated.View>
         </View>
-        <NextButton activateNext={activateNext} onPress={goToLocationForm} />
+        <NextButton disable={disableNext} onPress={goToLocationForm} />
       </SafeAreaView>
     </LinearGradient>
   );
